+++ src/client/src/components/alert/alertBanner.tsx
@@ -11,7 +11,7 @@ export const AlertBanner = () => {
         anchorOrigin={{ vertical: "top", horizontal: "center" }}
         autoHideDuration={alertContext.autoHideDuration}
         onClose={alertContext.closeAlert}>
-        <Alert variant="filled" severity={alertContext.severity} onClose={alertContext.closeAlert}>
+        <Alert variant={alertContext.variant} severity={alertContext.severity} onClose={alertContext.closeAlert}>
           {alertContext.text}
         </Alert>
       </Snackbar>