+++ src/client/src/components/alert/alertManager.tsx
@@ -1,15 +1,23 @@
 import { useCallback, useEffect, useState } from "react";
 import { AlertColor } from "@mui/material";
-import { AlertOptions } from "./alertInterfaces";
+import { AlertOptions, AlertVariant } from "./alertInterfaces";
 
 export const useAlertManager = () => {
   const [currentAlert, setCurrentAlert] = useState<AlertOptions>();
   const [alerts, setAlerts] = useState<AlertOptions[]>([]);
 
-  const showAlert = useCallback((text: string, severity: AlertColor | undefined, allowAutoHide?: boolean) => {
-    const newAlert = { text, severity: severity ?? "info", allowAutoHide: allowAutoHide ?? true };
-    setAlerts(prevAlerts => [...prevAlerts, newAlert]);
-  }, []);
+  const showAlert = useCallback(
+    (text: string, severity: AlertColor | undefined, allowAutoHide?: boolean, variant?: AlertVariant) => {
+      const newAlert = {
+        text,
+        severity: severity ?? "info",
+        allowAutoHide: allowAutoHide ?? true,
+        variant: variant ?? "filled",
+      };
+      setAlerts(prevAlerts => [...prevAlerts, newAlert]);
+    },
+    [],
+  );
 
   const closeAlert = () => {
     setCurrentAlert(undefined);
@@ -26,6 +34,7 @@ export const useAlertManager = () => {
     alertIsOpen: currentAlert?.text != null,
     text: currentAlert?.text,
     severity: currentAlert?.severity,
+    variant: currentAlert?.variant,
     autoHideDuration: currentAlert?.allowAutoHide === true ? 6000 : null,
     showAlert,
     closeAlert,