+++ src/client/src/pages/detail/form/stratigraphy/stratigraphy.ts
@@ -1,6 +1,5 @@
 import { UseFormReturn } from "react-hook-form";
 import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query";
-import { NullableDateString } from "../../../../api/apiInterfaces.ts";
 import { boreholeQueryKey } from "../../../../api/borehole.ts";
 import { ExtractionBoundingBox } from "../../../../api/dataextractionInterfaces.ts";
 import { fetchApiV2WithApiError } from "../../../../api/fetchApiV2.ts";
@@ -13,6 +12,7 @@ import {
   Stratigraphy,
   User,
 } from "../../../../api/generated";
+import { NullableDateString } from "../../../../api/unionTypes.ts";
 import { useResetTabStatus } from "../../../../hooks/useResetTabStatus.ts";
 
 export type DescriptionKind = "lithological" | "facies";
@@ -54,14 +54,6 @@ export type Lithology = BaseLayer &
 
 export interface LithologyFormValues extends Lithology {
   lithologicalDescription?: { description?: string };
-}
-
-export interface LayerDepth {
-  fromDepth: number;
-  toDepth: number;
-  lithologyId: number;
-  hasFromDepthError?: boolean;
-  hasToDepthError?: boolean;
 }
 
 export interface LithologyEditForm {