+++ src/client/src/pages/settings/admin/dialogs/entityType.ts
@@ -1,0 +1,4 @@
+export enum EntityType {
+  user = "User",
+  workgroup = "Workgroup",
+}